envsetup: set OUT_DIR to an absolute path always
OUT_DIR was set to $(TOPDIR)out previously,
but $(TOPDIR) was null, so it was a relative path.
This broke releasetools, inline kernel building, etc
since they require absolute paths.
Fix it so that it is set to $(shell readlink -f .)/out
if $(TOPDIR) is null.

Also remove hacks which checked if (OUT_DIR) was out
and changed it to $(ANDROID_BUILD_TOP)/out to workaround
the aforementioned problem.

build: work around missing readlink -f on Mac

envsetup: use $(CURDIR) for getting current directory

Change-Id: I6856b5ca265ea5de9dbc8a6675783360475ca8f5
Signed-off-by: Shekhawat2 <anandsingh215@yahoo.com>