build: kernel: Use LLVM_PREBUILTS_VERSION if no version is specified
Some ROMs may want to track Google's master branch, which has Clang 6
and 7 toolchains available, as it gets rid of some hacks in the kernel
source. Unfortunately, with the current default choosing logic, the
latest one would be chosen, which it shouldn't be. Use the variable set
in build/make/core/clang/versions.mk, as that's what was intended.

Signed-off-by: Shekhawat2 <anandsingh215@yahoo.com>